+++ tests/test_servicioPostal_Primera.cpp
@@ -11,8 +11,8 @@ namespace
         envioServicioPostalPrimera servicioPostal(3, 1);
 
         // Act - ejecute la operación
-        int actual = servicioPostal.calcular();
-        int esperada = 0.300;
+        double actual = servicioPostal.calcular();
+        double esperada = 0.300;
 
         // Assert - valide los resultados
         EXPECT_EQ(esperada, actual);
@@ -26,8 +26,8 @@ namespace
         envioServicioPostalPrimera servicioPostal(6, 1);
 
         // Act - ejecute la operación
-        int actual = servicioPostal.calcular();
-        int esperada = 0.450;
+        double actual = servicioPostal.calcular();
+        double esperada = 0.450;
 
         // Assert - valide los resultados
         EXPECT_EQ(esperada, actual);
@@ -41,8 +41,8 @@ namespace
         envioServicioPostalPrimera servicioPostal(9, 1);
 
         // Act - ejecute la operación
-        int actual = servicioPostal.calcular();
-        int esperada = 0.600;
+        double actual = servicioPostal.calcular();
+        double esperada = 0.600;
 
         // Assert - valide los resultados
         EXPECT_EQ(esperada, actual);
