+++ tests/test_servicioPostal_Segunda.cpp
@@ -11,8 +11,8 @@ namespace
         envioServicioPostalSegunda servicioPostal(3, 1);
 
         // Act - ejecute la operación
-        int actual = servicioPostal.calcular();
-        int esperada = 0.0280;
+        double actual = servicioPostal.calcular();
+        double esperada = 0.0280;
 
         // Assert - valide los resultados
         EXPECT_EQ(esperada, actual);
@@ -26,8 +26,8 @@ namespace
         envioServicioPostalSegunda servicioPostal(6, 1);
 
         // Act - ejecute la operación
-        int actual = servicioPostal.calcular();
-        int esperada = 0.0530;
+        double actual = servicioPostal.calcular();
+        double esperada = 0.0530;
 
         // Assert - valide los resultados
         EXPECT_EQ(esperada, actual);
@@ -41,8 +41,8 @@ namespace
         envioServicioPostalSegunda servicioPostal(9, 1);
 
         // Act - ejecute la operación
-        int actual = servicioPostal.calcular();
-        int esperada = 0.0750;
+        double actual = servicioPostal.calcular();
+        double esperada = 0.0750;
 
         // Assert - valide los resultados
         EXPECT_EQ(esperada, actual);
